requirements.txt
flask
flask-cors
python-dotenv
schedule
pymodbus
